+++ src/components/Main/Selector/SelectorBuilder/BuilderButton/BuilderButton.styles.jsx
@@ -11,22 +11,22 @@ export const $BuilderButton = styled.div`
   text-align: center;
   display: flex;
   align-items: center;
-  color: ${({ showCriteria }) => showCriteria !== "" ? appwhite : appgrey};
-  background: ${({ showCriteria }) => showCriteria !== "" ? appgreen : ""};
-  cursor: ${({ showCriteria }) => showCriteria !== "" ? 'pointer' : ''};
-  border: ${({ showCriteria }) => showCriteria !== "" ? `1px solid ${appgreen}` : `1px solid ${appgrey}`};
+  color: ${({ $showCriteria }) => $showCriteria !== "" ? appwhite : appgrey};
+  background: ${({ $showCriteria }) => $showCriteria !== "" ? appgreen : ""};
+  cursor: ${({ $showCriteria }) => $showCriteria !== "" ? 'pointer' : ''};
+  border: ${({ $showCriteria }) => $showCriteria !== "" ? `1px solid ${appgreen}` : `1px solid ${appgrey}`};
   user-select: none;
   padding: 5px 8px;
   transition: 0.5s;
 
   &:hover {
-    background: ${({ showCriteria }) => showCriteria !== "" ? appgreenlight : ""};
-    border: ${({ showCriteria }) => showCriteria !== "" ? `1px solid ${appgreenlight}` : `1px solid ${appgrey}`};
+    background: ${({ $showCriteria }) => $showCriteria !== "" ? appgreenlight : ""};
+    border: ${({ $showCriteria }) => $showCriteria !== "" ? `1px solid ${appgreenlight}` : `1px solid ${appgrey}`};
   }
 
   &:active {
-    color: ${({ showCriteria }) => showCriteria !== "" ? appgreen : appgrey};
-    border: ${({ showCriteria }) => showCriteria !== "" ? `1px solid ${appgreenlight}` : `1px solid ${appgrey}`};
+    color: ${({ $showCriteria }) => $showCriteria !== "" ? appgreen : appgrey};
+    border: ${({ $showCriteria }) => $showCriteria !== "" ? `1px solid ${appgreenlight}` : `1px solid ${appgrey}`};
   }
 
   & div {