+++ src/components/Main/Selector/SelectorBuilder/SelectorBuilder.jsx
@@ -2,6 +2,7 @@ import $SelectorBuilder from "./SelectorBuilder.styles.jsx";
 import BuilderButton from "./BuilderButton/BuilderButton.jsx";
 import BuilderFilter from "./BuilderFilter/BuilderFilter.jsx";
 import BuilderLabel from "./BuilderLabel/BuilderLabel.jsx";
+import BuilderValues from "./BuilderValues/BuilderValues.jsx";
 
 export default function SelectorBuilder() {
   return (
@@ -10,6 +11,7 @@ export default function SelectorBuilder() {
       <BuilderFilter criteria='show'/>
       <BuilderLabel />
       <BuilderFilter criteria='where'/>
+      <BuilderValues />
     </$SelectorBuilder>
   );
 }