+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add Metrics and Selector nav items routing
 * Add loaders for Metrics and Selector routes
 * Create and implement context and reducer for selector
+* Implement new api call for filters and response handler
 
 0.20 (2024-06-25)
 ===============